:recycle: REFACTOR: use absolute paths + move types to single folder
