+++ lib/initializers/index.ts
@@ -1,6 +1,6 @@
+import { GoodChatConfig } from 'lib/typings/goodchat';
 import _                  from 'lodash'
 import requireDir         from 'require-dir'
-import { GoodChatConfig } from '../types'
 import { each }           from '../utils/async'
 
 /**
