+++ lib/middlewares/rest/index.ts
@@ -1,5 +1,5 @@
 import Router             from '@koa/router'
-import { GoodChatConfig } from '../../types';
+import { GoodChatConfig } from 'lib/typings/goodchat';
 
 /**
  * Creates all the rest endpoints to read/write on conversations
