+++ lib/middlewares/webhooks/index.ts
@@ -1,13 +1,10 @@
-import Router               from '@koa/router'
-import { GoodChatConfig }   from '../../types'
-import logger               from '../../utils/logger'
-import { setupWebhooks }    from './setup'
-import { each }             from '../../../lib/utils/async'
-import { IntegrationsApi }  from 'sunshine-conversations-client'
-import {
-  WebhookEventBase,
-  WebhookPayload
-} from './typing'
+import Router                               from '@koa/router'
+import logger                               from 'lib/utils/logger'
+import { setupWebhooks }                    from './setup'
+import { each }                             from 'lib/utils/async'
+import { IntegrationsApi }                  from 'sunshine-conversations-client'
+import { GoodChatConfig }                   from 'lib/typings/goodchat'
+import { WebhookEventBase, WebhookPayload } from 'lib/typings/webhook_types'
 
 const { info } = logger('webhooks');
