+++ lib/services/events.ts
@@ -1,17 +1,8 @@
-import _                      from "lodash"
-import EventEmitter           from "events"
-
-import {
-  initializeCustomer,
-  sunshineUserToCustomer
-} from "../../lib/functions/customers"
-
-import { initializeConversation } from "../../lib/functions/conversations"
-
-import {
-  ConversationCreatedEvent,
-  WebhookEventBase
-} from "../middlewares/webhooks/typing"
+import _                                              from "lodash"
+import EventEmitter                                   from "events"
+import { initializeCustomer, sunshineUserToCustomer } from "./customers"
+import { initializeConversation }                     from "./conversations"
+import { ConversationCreatedEvent, WebhookEventBase } from "lib/typings/webhook_types";
 
 const emitter = new EventEmitter();
