+++ package.json
@@ -9,8 +9,8 @@
     "postbuild": "npm run copy:assets",
     "staging": "NODE_ENV=staging node dist/bin/serve.js",
     "production": "NODE_ENV=production node dist/bin/serve.js",
-    "dev": "NODE_ENV=development ts-node-dev bin/serve.ts",
-    "test": "NODE_ENV=test kankyo exec -- mocha -r ts-node/register --file specs/spec_helpers/setup.ts 'specs/**/*.spec.ts'",
+    "dev": "NODE_ENV=development ts-node-dev -r tsconfig-paths/register bin/serve.ts",
+    "test": "NODE_ENV=test kankyo exec -- mocha -r ts-node/register -r tsconfig-paths/register --file specs/spec_helpers/setup.ts 'specs/**/*.spec.ts'",
     "test:fast": "npm run test -- --bail",
     "test:coverage": "nyc npm run test",
     "coverage:json": "nyc report --reporter=json",
@@ -74,8 +74,9 @@
     "supertest-koa-agent": "^0.3.2",
     "ts-node-dev": "^1.0.0",
     "tsc-watch": "^4.2.9",
+    "tsconfig-paths": "^3.9.0",
     "typedoc": "^0.20.18",
-    "typescript": "^4.0.3"
+    "typescript": "^4.2.2"
   },
   "dependencies": {
     "@koa/router": "^9.4.0",
