+++ tsconfig.json
@@ -1,9 +1,8 @@
 {
   "include": [
-    "typings",
-    "lib",
-    "bin",
-    "index.ts"
+    "./lib",
+    "./bin",
+    "./index.ts"
   ],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
@@ -48,15 +47,15 @@
 
     /* Module Resolution Options */
     "moduleResolution": "node",               /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */
-    "baseUrl": "./",                          /* Base directory to resolve non-absolute module names. */
+    "baseUrl": ".",                          /* Base directory to resolve non-absolute module names. */
     "paths": {
       "lib/*": [
         "./lib/*"
       ],
     },                                        /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
-    "typeRoots": ["./typings"],                 /* List of folders to include type definitions from. */
-    "types": ["node"],                        /* Type declaration files to be included in compilation. */
+    // "typeRoots": [],                          /* List of folders to include type definitions from. */
+    // "types": [],                              /* Type declaration files to be included in compilation. */
     // "allowSyntheticDefaultImports": true,  /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
     "esModuleInterop": true,                  /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
     // "preserveSymlinks": true,              /* Do not resolve the real path of symlinks. */
